Reading from recipe from file + more rest recipes

+++ src/main/kotlin/beerxml/MashStep.kt
@@ -8,9 +8,9 @@ data class MashStep(
     @JacksonXmlProperty(localName = "TYPE") val type: String?, // TYPE
     @JacksonXmlProperty(localName = "INFUSE_AMOUNT") val infuseAmount: Double?, // INFUSE_AMOUNT
     @JacksonXmlProperty(localName = "STEP_TIME") val stepTime: Int?, // STEP_TIME
-    @JacksonXmlProperty(localName = "STEP_TEMP") val stepTemp: Int?, // STEP_TEMP
+    @JacksonXmlProperty(localName = "STEP_TEMP") val stepTemp: Double?, // STEP_TEMP
     @JacksonXmlProperty(localName = "RAMP_TIME") val rampTime: Int?, // RAMP_TIME
-    @JacksonXmlProperty(localName = "END_TEMP") val endTemp: Int?, // END_TEMP
+    @JacksonXmlProperty(localName = "END_TEMP") val endTemp: Double?, // END_TEMP
     @JacksonXmlProperty(localName = "DESCRIPTION") val description: String?, // DESCRIPTION
     @JacksonXmlProperty(localName = "WATER_GRAIN_RATIO") val waterGrainRatio: Double?, // WATER_GRAIN_RATIO
     @JacksonXmlProperty(localName = "DECOCTION_AMT") val decoctionAmount: Double?, // DECOCTION_AMT
